Fixed up surveys to load in the same way
