+++ src/decide.ts
@@ -70,22 +70,7 @@ export class Decide {
 
         this.instance._afterDecideResponse(response)
 
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        const surveysGenerator = window?.extendPostHogWithSurveys
-
         // TODO: Need to change this to be able to be more reactive 
-        if (!this.instance.config.disable_surveys && response['surveys'] && !surveysGenerator) {
-            loadScript(this.instance.requestRouter.endpointFor('assets', '/static/surveys.js'), (err) => {
-                if (err) {
-                    return logger.error(`Could not load surveys script`, err)
-                }
-
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-ignore
-                window.extendPostHogWithSurveys(this.instance)
-            })
-        }
 
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
